refactor : 컨트롤러에 있던 로또 결과 생성 비즈니스 로직을 도메인 객체로 이동
