+++ src/main/java/lotto/domain/ticket/LottoTickets.java
@@ -1,7 +1,13 @@
 package lotto.domain.ticket;
 
+import lotto.domain.LottoResult;
+import lotto.domain.Money;
+
 import java.util.Collections;
 import java.util.List;
+import java.util.function.Function;
+
+import static java.util.stream.Collectors.*;
 
 public class LottoTickets {
     private final List<LottoTicket> lottoTickets;
@@ -21,5 +27,13 @@ public class LottoTickets {
     public LottoTickets concat(LottoTickets that) {
         this.lottoTickets.addAll(that.list());
         return new LottoTickets(lottoTickets);
+    }
+
+    public LottoResult calculateLottoResult(WinningLottoTicket winningLottoTicket, Money lottoPrice) {
+        return lottoTickets.stream()
+                .map(winningLottoTicket::compareNumbers)
+                .collect(collectingAndThen(
+                        groupingBy(Function.identity(), counting()),
+                        lottoResultMap -> new LottoResult(lottoResultMap, lottoPrice.multiply(lottoTickets.size()))));
     }
 }
